% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalinf.R
\name{estimate_causal_effect}
\alias{estimate_causal_effect}
\title{Estimate causal effect of a single variable using the backdoor criterion}
\usage{
estimate_causal_effect(data, x, y, confounders, binary_vars)
}
\arguments{
\item{data}{A data frame containing the variables of interest}

\item{x}{The name of the treatment variable}

\item{y}{The name of the outcome variable}

\item{confounders}{A vector of names of confounding variables}

\item{binary_vars}{A vector of names of binary variables}
}
\value{
An estimate of the causal effect of x on y
}
\description{
Estimate causal effect of a single variable using the backdoor criterion
}
