% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalinf.R
\name{estimate_causal_effects}
\alias{estimate_causal_effects}
\title{Estimate causal effects of multiple variables using the backdoor criterion}
\usage{
estimate_causal_effects(dataset, treatment_var, outcome_var, alpha = 0.05)
}
\arguments{
\item{dataset}{A data frame containing the variables of interest}

\item{treatment_var}{A vector of names of the treatment variables}

\item{outcome_var}{The name of the outcome variable}

\item{alpha}{The significance value used}
}
\value{
A list of estimates of the causal effects of each treatment variable on the outcome
}
\description{
Estimate causal effects of multiple variables using the backdoor criterion
}
